# Remove get_insn_idx from iseq_inline_constant_cache struct 