--- a/vm_core.h
+++ b/vm_core.h
@@ -269,9 +269,6 @@ set_ic_serial(struct iseq_inline_constant_cache_entry *ice, rb_serial_t v)
 
 struct iseq_inline_constant_cache {
     struct iseq_inline_constant_cache_entry *entry;
-    // For YJIT: the index to the opt_getinlinecache instruction in the same iseq.
-    // It's set during compile time and constant once set.
-    unsigned get_insn_idx;
 };
 
 struct iseq_inline_iv_cache_entry {