--- a/yjit_iface.c
+++ b/yjit_iface.c
@@ -600,6 +600,29 @@ rb_yjit_constant_state_changed(void)
     }
 }
 
+static unsigned
+find_get_insn_idx(const rb_iseq_t *const iseq, IC ic) {
+    const struct rb_iseq_constant_body *const body = iseq->body;
+    VALUE *code = body->iseq_encoded;
+
+    // This should come from a running iseq, so direct threading translation
+    // should have been done
+    RUBY_ASSERT(FL_TEST((VALUE)iseq, ISEQ_TRANSLATED));
+
+    for (unsigned idx = 0; idx < body->iseq_size; ) {
+        VALUE insn = rb_vm_insn_addr2insn((void *)code[idx]);
+
+        // opt_getinlinecache OFFSET, IC
+        if (insn == BIN(opt_getinlinecache) && (IC)code[idx + 2] == ic) {
+            return idx;
+        }
+
+        idx += insn_len(insn);
+    }
+
+    rb_bug("yjit: could not find matching opt_getinlinecache");
+}
+
 // Callback from the opt_setinlinecache instruction in the interpreter.
 // Invalidate the block for the matching opt_getinlinecache so it could regenerate code
 // using the new value in the constant cache.
@@ -619,42 +642,37 @@ rb_yjit_constant_ic_update(const rb_iseq_t *const iseq, IC ic)
     {
         const struct rb_iseq_constant_body *const body = iseq->body;
         VALUE *code = body->iseq_encoded;
-        const unsigned get_insn_idx = ic->get_insn_idx;
+        const unsigned get_insn_idx = find_get_insn_idx(iseq, ic);
 
-        // This should come from a running iseq, so direct threading translation
-        // should have been done
-        RUBY_ASSERT(FL_TEST((VALUE)iseq, ISEQ_TRANSLATED));
-        RUBY_ASSERT(get_insn_idx < body->iseq_size);
+        RUBY_ASSERT(get_insn_idx == ic->get_insn_idx);
+
         RUBY_ASSERT(rb_vm_insn_addr2insn((const void *)code[get_insn_idx]) == BIN(opt_getinlinecache));
+        RUBY_ASSERT(get_insn_idx < body->iseq_size);
+        RUBY_ASSERT(insn_op_type(BIN(opt_getinlinecache), 1) == TS_IC);
+        RUBY_ASSERT(ic == (IC)code[get_insn_idx + 1 + 1]);
 
         // Find the matching opt_getinlinecache and invalidate all the blocks there
-        RUBY_ASSERT(insn_op_type(BIN(opt_getinlinecache), 1) == TS_IC);
-        if (ic == (IC)code[get_insn_idx + 1 + 1]) {
-            rb_yjit_block_array_t getinlinecache_blocks = yjit_get_version_array(iseq, get_insn_idx);
+        rb_yjit_block_array_t getinlinecache_blocks = yjit_get_version_array(iseq, get_insn_idx);
 
-            // Put a bound for loop below to be defensive
-            const size_t initial_version_count = rb_darray_size(getinlinecache_blocks);
-            for (size_t iteration=0; iteration<initial_version_count; ++iteration) {
-                getinlinecache_blocks = yjit_get_version_array(iseq, get_insn_idx);
+        // Put a bound for loop below to be defensive
+        const size_t initial_version_count = rb_darray_size(getinlinecache_blocks);
+        for (size_t iteration=0; iteration<initial_version_count; ++iteration) {
+            getinlinecache_blocks = yjit_get_version_array(iseq, get_insn_idx);
 
-                if (rb_darray_size(getinlinecache_blocks) > 0) {
-                    block_t *block = rb_darray_get(getinlinecache_blocks, 0);
-                    invalidate_block_version(block);
+            if (rb_darray_size(getinlinecache_blocks) > 0) {
+                block_t *block = rb_darray_get(getinlinecache_blocks, 0);
+                invalidate_block_version(block);
 #if YJIT_STATS
-                    yjit_runtime_counters.invalidate_constant_ic_fill++;
+                yjit_runtime_counters.invalidate_constant_ic_fill++;
 #endif
-                }
-                else {
-                    break;
-                }
             }
+            else {
+                break;
+            }
+        }
 
-            // All versions at get_insn_idx should now be gone
-            RUBY_ASSERT(0 == rb_darray_size(yjit_get_version_array(iseq, get_insn_idx)));
-        }
-        else {
-            RUBY_ASSERT(false && "ic->get_insn_diex not set properly");
-        }
+        // All versions at get_insn_idx should now be gone
+        RUBY_ASSERT(0 == rb_darray_size(yjit_get_version_array(iseq, get_insn_idx)));
     }
     RB_VM_LOCK_LEAVE();
 }