I want to reduce the size of the `iseq_inline_storage_entry` union, which is currently 16 bytes due to the `get_insn_idx` used by YJIT and because of the `once` struct (which I'll tackle in a separate PR).

```
union iseq_inline_storage_entry {
        struct {
                struct rb_thread_struct * running_thread;                                /*     0     8 */
                /* typedef VALUE */ long unsigned int  value;                            /*     8     8 */
        } once;                                                                          /*     0    16 */
        struct iseq_inline_constant_cache {
                struct iseq_inline_constant_cache_entry * entry;                         /*     0     8 */
                unsigned int       get_insn_idx;                                         /*     8     4 */
        }ic_cache; /*     0    16 */
        struct iseq_inline_iv_cache_entry {
                struct rb_iv_index_tbl_entry * entry;                                    /*     0     8 */
        }iv_cache; /*     0     8 */
};
```

This PR replaces the stored index with a scan to find a matching instruction and operand.

This is slower than just storing the index, but I think we should expect IC updates to be uncommon and tolerate them being slow. This would also be made faster by #5622 if merged.

This is similar to what is done in rb_iseq_mark for direct marking.