@maximecb 2022-03-03
I'd like to get a review from @XrXr on this, but a priori: is the small size advantage worth the added complexity and potential slowdown?

In terms of constant IC updates being slow, in SFR we're seeing on the order of 10-20K IIRC. Iterating over the entire iseq body seems problematic.

Lastly @kddnewton has a PR touching the constant IC code, trying to improve the constant IC invalidation mechanism, and we're in the process of porting YJIT to Rust, so we might prefer to keep this code stable for now if possible 🙏