@XrXr 2022-03-03
Constant expression ICs busting is dominant cause of invalidation in our workloads both during boot and at runtime. This is partly why @kddnewton is looking to reducing these IC busts. Storing the index is also informed by these workloads so I think scanning `encoded_iseq` is going to be prohibitive for at least warmup perf.