@jhawthorn 2022-03-03
> is the small size advantage worth the added complexity and potential slowdown

IMO this is a reduction in complexity since we no longer need to include hacks for bookkeeping during iseq compilation.

As for the size advantage, I do think it's important. ISEQs are extremely common and in our app account for the vast majority of our memory use. I don't know how much is caused by ISEs (one for every constant, ivar, or cvar reference), but doubling their size is not ideal, neither is spreading them across more cache lines.

Removing the complexity of this field (and of the `once` struct) I believe is also necessary for further simplification and optimizations: since it will make all 

It also feels wrong that those not enabling YJIT have to pay the cost of this extra field. (Though someday I hope everyone is enabling YJIT 🙂)

> Lastly @kddnewton has a PR touching the constant IC code, trying to improve the constant IC invalidation mechanism

From a quick scan there are no conflicts between these approaches, and perhaps that would reduce the number of invalidations.