@kddnewton 2022-03-03
I don't think this code is in conflict with my PR.

That being said, I would echo what Alan has already said about it being a potential issue with our workloads since we bust caches so much. I'd imagine @casperisfine would have some more insight here since he's been doing a ton of benchmarking/work on this area lately.