@maximecb 2022-03-03
> As for the size advantage, I do think it's important. ISEQs are extremely common and in our app account for the vast majority of our memory use. 
> It also feels wrong that those not enabling YJIT have to pay the cost of this extra field. (Though someday I hope everyone is enabling YJIT 🙂)

Your other PR #5622 is going to increase memory usage much more than this tiny reduction in memory usage though.

Scanning the iseq body on each invalidation seems not great. The constant cache state gets bumped a lot, and each time it gets bumped, it invalidates a lot of ICs. If each of these has to scan through the iseq, it could be pretty bad.