@XrXr 2022-03-03
I think you might be overestimating the memory savings. A while ago I was able to [cut 8 bytes from each call instruction](https://bugs.ruby-lang.org/issues/16258), and that amounted to a meager amount of memory savings in our app with LOC in the millions. Calls are probably more common than entries in this array of unions so this PR is going to provide less savings.

If you really want this memory saving, you could wrap the struct field and relevant code in `compile.c` with `#if USE_MJIT && defined(YJIT_SUPPORTED_P)` that way interpreter only builds can have different trade offs. Rust YJIT is going to require explicit `configure` opt-in for a little while a least, so the field _not_ existing is going to be default.