@jhawthorn 2022-03-03
> Scanning the iseq body on each invalidation seems not great. The constant cache state gets bumped a lot, and each time it gets bumped, it invalidates a lot of ICs. If each of these has to scan through the iseq, it could be pretty bad.

Fortunately that's not what happens. The constant state being bumped invalidates all the blocks in an `st_foreach`. This is only hit when we set the IC and we only scan for one then.

> If you really want this memory saving, you could wrap the struct field and relevant code in compile.c with #if USE_MJIT && defined(YJIT_SUPPORTED_P) that way interpreter only builds can have different trade offs

The ifdef approach doesn't accomplish what I want. My primary interest is a simplification in how these are represented. Also, I hope to be enabling YJIT.

---

Do you have a benchmark of warmup you'd like to test this against? I'm still convinced the existing representation should to be removed from that struct, if performance of that lookup really is that important there are other places we can store that index.