@jhawthorn 2022-03-03
Here's the most IC-heavy benchmark I could think of:

```
N = 1000
class Foo
  BAR = "bar"
  N.times do |i|
    class_eval "def bar_#{i}; 1+1; BAR; end"
  end

  class_eval "def all_bar; #{N.times.map{|i| "bar_#{i};" }.join}; end"
end

M = 1000
M.times do |i|
  Foo.new.all_bar

  1.times do
    Object.const_set(:BAZ, i)
    Object.send(:remove_const, :BAZ)
  end
end
```

Before: 
```
./miniruby --yjit benchmark_ic_fills.rb  1.55s user 6.30s system 99% cpu 7.851 total
```

After:
```
./miniruby --yjit benchmark_ic_fills.rb  1.68s user 6.29s system 99% cpu 7.974 total
```

To me this result sounds right. I'd expect generating and invalidating the bytecode to be much more significant than a read through the iseq, but maybe there's a benchmark that shows something different?