@XrXr 2022-03-03
Iseq size is proportional to the number of lines in the method, also the scan happens per constant expression in the method when they are ran in the interpreter so I think micro benchmarks are not going to be very helpful to evaluate this as perf impact is related to the population of iseqs in the app. I'll find some time to try this on our monolith next week. I'd expect an impact to boot performance.