@jhawthorn 2022-03-04
Thanks Alan. Looking forward to the results 🙇‍♂️

If there does end up being a performance regression, knowing the source would be helpful as I do think there will be a good way to make this work.

A couple ideas:
* We could add a flag to the const cache IMEMO indicating when it's getinlinecache insn has been jitted, so we don't do any work on un-jitted code (I like this and it might be nice to do anyways)
* On 64-bit platforms, there's a spare VALUE on the const cache IMEMO we could make use of
* There's probably some way to reduce the part of the iseq we search for (more complex though)
  * Possibly the `insns_info` struct could help, we probably have some guarantee about the set/getinlinecache matching here
  * I like this less than the other two ideas